requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==6.1.0
inflection==0.5.1
mccabe==0.7.0
packaging==23.2
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
